requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
discord.py==2.2.3
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
msgpack==1.0.5
multidict==6.0.4
requests==2.31.0
urllib3==2.0.3
yarl==1.9.2
